Fix automatic push to account in Login VC
presentWithSegueIdentifier(_:animated:) was being called in both
viewWillAppear and viewDidAppear, which had the effect of bouncing the
user back to the account VC whenever they hit "back". These were removed
in favour of one call in viewDidLoad.

This will still trigger the navigation if the view is dropped for
end-user-opaque reasons. It might be better to add state to this VC to
track "first launch", and only navigate in that case, but this might
muddy the sample code too much.